+++ user/src/components/user/taskdeal.vue
@@ -66,8 +66,8 @@
       <el-button type="primary" @click="dialogFormVisible1 = true">取消任务</el-button>
 
 <el-dialog title="取消任务" :visible.sync="dialogFormVisible1">
-  <el-form :model="upform1">
-    <el-form-item label="理由类型">
+  <el-form :model="upform1" ref="upform1" :rules="rules">
+    <el-form-item label="理由类型" prop="type">
     <el-select v-model="upform1.type" placeholder="请选择理由类型" >
       <el-option label="我已完成" value="1"></el-option>
       <el-option label="太久无人接单" value="2"></el-option>
@@ -76,28 +76,28 @@
       <el-option label="其他" value="5"></el-option>
     </el-select>
   </el-form-item>
-    <el-form-item label="描述" :label-width="formLabelWidth">
+    <el-form-item label="描述" :label-width="formLabelWidth" prop="content">
       <el-input type="textarea" v-model="upform1.content" autocomplete="off"></el-input>
     </el-form-item>
   </el-form>
   <div slot="footer" class="dialog-footer">
     <el-button @click="dialogFormVisible1 = false">取 消</el-button>
-    <el-button type="primary" @click="handleCancel()">确 定</el-button>
+    <el-button type="primary" @click="handleCancel('upform1')">确 定</el-button>
   </div>
 </el-dialog>
       <br /><br />
       <el-button type="primary"  @click="dialogFormVisible2 = true"> 欺诈反馈</el-button>
 
 <el-dialog title="任务反馈" :visible.sync="dialogFormVisible2">
-  <el-form :model="upform2">
-    <el-form-item label="理由" :label-width="formLabelWidth">
+  <el-form :model="upform2" ref="upform2" :rules="rules2">
+    <el-form-item label="理由" :label-width="formLabelWidth" prop="content">
       <el-input type="textarea" v-model="upform2.content" autocomplete="off"></el-input>
     </el-form-item>
 
   </el-form>
   <div slot="footer" class="dialog-footer">
     <el-button @click="dialogFormVisible2 = false">取 消</el-button>
-    <el-button type="primary" @click="handleBack()">确 定</el-button>
+    <el-button type="primary" @click="handleBack('upform2')">确 定</el-button>
   </div>
 </el-dialog>
       <br /><br />
@@ -167,6 +167,20 @@ export default {
         dialogFormVisible1:false,
         dialogFormVisible2:false,
         tid:{id:''},
+         rules: {
+          type: [
+            { required: true, message: '请选择理由类型', trigger: 'change' }
+          ],
+          content: [
+            { required: true, message: '请输入描述', trigger: 'blur' }
+          ]
+        },
+        rules2: {
+          
+          content: [
+            { required: true, message: '请输入反馈描述', trigger: 'blur' }
+          ]
+        }
     }
    
    
@@ -195,7 +209,9 @@ export default {
      var sendJson = this.upform1;
      console.log(sendJson);
       var self=this;
-        axios.delete('/task/receiver/cancel/'+self.detail.id,{data:sendJson})
+      this.$refs[formName].validate((valid) => {
+          if (valid) {
+            axios.delete('/task/receiver/cancel/'+self.detail.id,{data:sendJson})
            .then(response => {
              
               if(response.status=="200")
@@ -213,6 +229,13 @@ export default {
         });
               }
             });
+            
+          } else {
+            console.log('error submit!!');
+            return false;
+          }
+        });
+        
     	this.dialogFormVisible1 = false;
     },
     //任务反馈
@@ -220,7 +243,9 @@ export default {
       this.upform2.task=this.detail.id;
      var sendJson = this.upform2;
       var self=this;
-        axios.post('/feedback/suitTask',sendJson)
+      this.$refs[formName].validate((valid) => {
+          if (valid) {
+             axios.post('/feedback/suitTask',sendJson)
            .then(response => {
              
               if(response.status=="200")
@@ -238,6 +263,13 @@ export default {
         });
               }
             });
+            
+          } else {
+            console.log('error submit!!');
+            return false;
+          }
+        });
+       
     	this.dialogFormVisible2 = false;
     },
     compelete(){
